Add dwm reload & remove config.h from repo
Have added a dwm launch script which will be triggered
from the .xinitrc. It will log out errors to the desired
location and will also restart dwm if I quit normally,
if I have updated my config or anything else. I have also
added a keypinding to call `pkill` and kill the dwm process
so I can exit when I need to.

The other thing of note was removing the config.h file this will
be followed by adding it to the gitignore to enforce that I update
my config through the config.def.h file as some patches are applied
to that and in theory they should always be the same.